refactor(console): move ConnectorVM to central location
