+++ gravitee-apim-console-webui/src/entities/management-api-v2/api/v4/index.ts
@@ -21,6 +21,7 @@ export * from './baseListener';
 export * from './baseSelector';
 export * from './channelSelector';
 export * from './conditionSelector';
+export * from './connectorVM';
 export * from './dlq';
 export * from './endpointGroupServices';
 export * from './endpointGroupV4';
